Patch NOVAS to allow flexible reference points for proper motion.
NOVAS's calculation of the effect of proper motion assumes that the
coordinates of the source are not only in the ICRS J2000 frame, but
that the source was at the given coordinates precisely at J2000.0.
This is not generally true if the source has nontrivial proper motion:
the reported position may be in the right frame, but there's still the
question of how much the source may have moved in the interval between
the observation and January 1.5, 2000.

My solution is to add another source parameter, 'promoepoch', which
specifies the TDB JD at which the effect of proper motion is defined
to be zero. This is easy to integrate into the existing computations.
When relevant it is set to T0, NOVAS's name for the JD of J2000.0,
which leaves the standard behavior unchanged. But it can be overridden
for those cases when you know not only the source's coordinates but
when they were measured.

An alternative solution would be to provide a utility routine to
adjust the stored RA and Dec coordinates given a known proper motion
and epoch of observation. But NOVAS' calculations are subtle and I
don't get the impression that this can easily be done safely. The
behavior of the proper_motion function is straightforward, on the
other hand, and it seems quite safe to offset the epoch as I've done.