compile_ephem.py: fix bug in compilation
It turns out we need to adjust the JD span records at the beginning
of the file. Otherwise everything breaks. That's lame because we can't
stream out the output, but oh well.

Fixing this gets my BJD calculation to agree with the OSU online
calculator to 0.2 s for the, uh, single test case I've tried.